<feature> add: Draw Circles

+++ src/GUI/GUI.c
@@ -5,11 +5,11 @@ bool GUI_Button(Rectangle rect, string text, int fontSize, Color btnColor, Color
     if (isHover) {
         DrawRectangleRec(rect, textColor);
         DrawRectangleLinesEx(rect, 1, textColor);
-        DrawText(text, rect.x + 10, rect.y + 10, fontSize, btnColor);
+        DrawText(text, (int)rect.x + 10, (int)rect.y + 10, fontSize, btnColor);
     } else {
         DrawRectangleRec(rect, btnColor);
         DrawRectangleLinesEx(rect, 1, textColor);
-        DrawText(text, rect.x + 10, rect.y + 10, fontSize, textColor);
+        DrawText(text, (int)rect.x + 10, (int)rect.y + 10, fontSize, textColor);
     }
     return isHover && IsMouseButtonPressed(MOUSE_LEFT_BUTTON);
 }
